refactored to remove duplicate classes, all currently implemented test cases pass.
