--- src/test/java/config.properties
@@ -1,8 +1,0 @@
-automationexcercise.base_uri=https://automationexercise.com/api/
-automationexcercise.api_path.get_all_brands=brandsList
-automationexcercise.api_path.get_all_products=productsList
-automationexcercise.api_path.get_all=brandsList
-default.email=testuser@example.com
-default.password=password123
-endpoint.createAccount=/createAccount
-endpoint.api_path.createAccount=/createAccount
