--- src/test/java/restassured/REST.java
@@ -1,15 +1,0 @@
-package restassured;
-
-import io.restassured.RestAssured;
-import io.restassured.response.Response;
-import org.hamcrest.MatcherAssert;
-import org.hamcrest.Matchers;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import utils.API;
-
-public class REST {
-
-
-}
