+++ src/test/resources/config.properties
@@ -5,7 +5,7 @@ endpoint.createAccount=/createAccount
 automationexercise.base_uri=https://automationexercise.com/api/
 automationexercise.api_path.get_all_brands=brandsList
 automationexercise.api_path.get_all_products=productsList
-automationexercise.api_path.get_all_=brandsList
+automationexercise.api_path.get_all=brandsList
 automationexercise.api_path.create_account=/createAccount
 automationexercise.api_path.post_all_products=productsList
 automationexercise.invalid_email=testex142638363t@example.com
